+++ CHANGES.rst
@@ -24,6 +24,7 @@ New features:
 
 Bug fixes:
 
+- Test that we can add an RRULE as a string. See `Issue 301 <https://github.com/collective/icalendar/issues/301>`_.
 - Test that we support dateutil timezones as outlined in `Issue 336 <https://github.com/collective/icalendar/issues/336>`_.